visualization/requirements.txt
flask
flask-socketio
kafka-python
eventlet
